requirements.txt
pyyaml==6.0.2
discord-py==2.4.0